PTT-11 Installing new library react-timer-hook for setting the timer. Combining word-checking logic and appropriate view generation for an event with a final styled component
