+++ package.json
@@ -19,6 +19,7 @@
     "react-dom": "^18.2.0",
     "react-router": "^6.3.0",
     "react-scripts": "5.0.1",
+    "react-timer-hook": "^3.0.5",
     "styled-components": "^5.3.5",
     "web-vitals": "^2.1.4",
     "yup": "^0.32.11"
