+++ src/App.js
@@ -1,7 +1,12 @@
 import React from "react";
+import { TypingTest } from "./components/TypingTest/TypingTest";
 
 const App = () => {
-  return <div></div>;
-
+  return (
+    <div>
+      <TypingTest />
+    </div>
+  );
+};
 
 export default App;
